refactor: move validateDefaultArgs to validate-default-args.mts
Keeps args.mts under the 200-line lint limit while making the shared
validation helper independently importable.
